+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/fab/MovableFloatingActionButton.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
+import expo.modules.devmenu.DevMenuLaunchOverrides
 import expo.modules.devmenu.compose.DevMenuState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
@@ -96,6 +97,7 @@ fun MovableFloatingActionButton(
     val fab = rememberFabState(bounds)
 
     val isFabDisplayable = state.showFab &&
+      DevMenuLaunchOverrides.canShowFab &&
       !state.isInPictureInPictureMode &&
       bounds.x >= 0f &&
       bounds.y >= 0f